pyproject.toml
[build-system]
requires = ["scikit-build-core >=0.4.3", "nanobind >=1.3.2"]
build-backend = "scikit_build_core.build"

[project]
name = "microseg"
version = "0.0.1"
description = "An example minimal project that compiles bindings using nanobind and scikit-build"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Anand Srinivasan", email = "" },
]
classifiers = [
    "License :: OSI Approved :: BSD License",
]
dependencies = [
    # GUI and visualization
    "qtpy",
    "PyQt6",
    "pyqtgraph >=0.13.7",
    "superqt",
    "pyopengl",
    "seaborn",
    "colorcet",
    
    # Image processing and computer vision
    "opencv-python",  # cv2
    "Pillow",        # PIL
    "pillow-heif",   # HEIC/HEIF support for Pillow
    "pillow-heif",
    "PyMuPDF",       # pymupdf
    "scikit-image",  # skimage
    "aicsimageio",
    "aicspylibczi>=3.1.1",
    "cellpose >=4.0.0",
    "tifffile",
    
    # Geometry and computation
    "numpy>=2.2.0",
    "scipy",
    "shapely",
    "upolygon",
    "shortuuid",
    "pywavelets",
    
    # Custom dependencies
    "matgeo",
    "asrvsn_math @ git+https://github.com/asrvsn/asrvsn_math.git"
]   

[project.urls]
Homepage = "https://github.com/asrvsn/microseg"


[tool.scikit-build]
# Protect the configuration against future changes in scikit-build-core
minimum-version = "0.4"

# Setuptools-style build caching in a local directory
build-dir = "build/{wheel_tag}"

# Build stable ABI wheels for CPython 3.12+
wheel.py-api = "cp312"

[tool.cibuildwheel]
# Necessary to see build output from the actual compilation
build-verbosity = 1

# Run pytest to ensure that the package was correctly built
test-command = "pytest {project}/tests"
test-requires = "pytest"

# Don't test Python 3.8 wheels on macOS/arm64
test-skip="cp38-macosx_*:arm64"

# Needed for full C++17 support
[tool.cibuildwheel.macos.environment]
MACOSX_DEPLOYMENT_TARGET = "10.14"
